double -> float. bc of SQLite

+++ src/main/java/com/zhilin/http/req/ReqOperation.java
@@ -10,7 +10,7 @@ public interface ReqOperation {
 
     Integer to();
 
-    double amount();
+    float amount();
 
     Type type();
